+++ tests/test_parameter_id.py
@@ -180,6 +180,7 @@ _SAMPLES: dict[str, ir.Feature] = {
     ),
     "ChamferFeature": ir.ChamferFeature(_F, "z", 2.0, 2.0, 45.0),
     "FilletFeature": ir.FilletFeature(_F, "z", 3.0),
+    "PairedRampStepFeature": ir.PairedRampStepFeature(_F, "z", 45.0, 12.0),
     "FlatFeature": ir.FlatFeature(_F, "z", 18.0),
     "GrooveFeature": ir.GrooveFeature(_F, "z", 3.0, 20.0),
     "StepLevelFeature": ir.StepLevelFeature(_F, 0.0, (5.0, 10.0, 15.0)),
@@ -261,6 +262,10 @@ _SAMPLE_BINDINGS = {
     "ExternalSpurGearFeature": (),
     "ChamferFeature": (("chamfer.length", 2.0),),
     "FilletFeature": (("fillet.radius", 3.0),),
+    "PairedRampStepFeature": (
+        ("ramp_angle.angle", 45.0),
+        ("ramp_run.length", 12.0),
+    ),
     "FlatFeature": (("flat.length", 18.0),),
     "GrooveFeature": (("groove.length", 3.0), ("groove.diameter", 20.0)),
     "StepLevelFeature": (